chore: moved mocks folder outside tests
