+++ packages/hooks/src/__tests__/useHasP2PSupportedCurrencies.spec.tsx
@@ -4,7 +4,7 @@ import { StoreProvider, mockStore } from '@deriv/stores';
 import { renderHook } from '@testing-library/react-hooks';
 import useHasP2PSupportedCurrencies from '../useHasP2PSupportedCurrencies';
 import type { TStores } from '@deriv/stores/types';
-import { withMockAPIProvider } from './mocks';
+import { withMockAPIProvider } from '../mocks';
 
 jest.mock('@deriv/api', () => ({
     ...jest.requireActual('@deriv/api'),
