+++ packages/hooks/src/__tests__/useVerifyEmail.spec.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { mockStore } from '@deriv/stores';
 import { act, renderHook } from '@testing-library/react-hooks';
 import useVerifyEmail from '../useVerifyEmail';
-import { withMockAPIProvider } from './mocks';
+import { withMockAPIProvider } from '../mocks';
 
 jest.mock('@deriv/api', () => ({
     ...jest.requireActual('@deriv/api'),
